Enable/Disable process launch by loc_launcher based on SOC-Id
Add a parameter SOC_IDS for process configuration in izat.conf
to include/exclude the process based on SOC-Id. Read the value
of this parameter to enable or disable the process launch by
loc_launcher.

Fixed an issue in read_a_line function to handle newline at the
end of line in the file to be read.

Change-Id: I66777856ebbb928ce4f49a44dfd9d2f1fbca69b5
CRs-Fixed: 2540349